refactor: remove legacy payment confirmation saga
