+++ src/main/java/kr/kro/airbob/domain/payment/event/PaymentEvent.java
@@ -1,6 +1,5 @@
 package kr.kro.airbob.domain.payment.event;
 
-import kr.kro.airbob.domain.payment.dto.PaymentRequest;
 import kr.kro.airbob.domain.payment.dto.TossPaymentResponse;
 import kr.kro.airbob.outbox.EventPayload;
 import lombok.AccessLevel;
@@ -8,20 +7,6 @@ import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class PaymentEvent {
-
-	public record PaymentCompletedEvent(String reservationUid) implements EventPayload {
-		@Override
-		public String getId() {
-			return this.reservationUid;
-		}
-	}
-
-	public record PaymentFailedEvent(String reservationUid, String reason) implements EventPayload{
-		@Override
-		public String getId() {
-			return this.reservationUid;
-		}
-	}
 
 	public record PaymentCancellationFailedEvent(String reservationUid, String reason) implements EventPayload{
 		@Override
@@ -44,28 +29,6 @@ public class PaymentEvent {
 	) implements EventPayload {
 		@Override
 		public String getId() { return reservationUid; }
-	}
-
-	public record PgCallSucceededEvent(
-		TossPaymentResponse response,
-		String reservationUid
-	) implements EventPayload {
-		@Override
-		public String getId() {
-			return this.reservationUid;
-		}
-	}
-
-	public record PgCallFailedEvent(
-		PaymentRequest.Confirm request, // 원래 요청 정보
-		String reservationUid,
-		String errorCode,
-		String errorMessage
-	) implements EventPayload {
-		@Override
-		public String getId() {
-			return this.reservationUid;
-		}
 	}
 
 	public record PgCancelCallSucceededEvent(
