+++ src/main/java/kr/kro/airbob/kafka/consumer/ReservationEventsConsumer.java
@@ -5,7 +5,6 @@ import org.springframework.kafka.support.Acknowledgment;
 import org.springframework.messaging.handler.annotation.Payload;
 import org.springframework.stereotype.Component;
 
-import kr.kro.airbob.domain.payment.event.PaymentEvent;
 import kr.kro.airbob.domain.reservation.event.ReservationEvent;
 import kr.kro.airbob.domain.reservation.service.ReservationHoldService;
 import kr.kro.airbob.domain.reservation.service.ReservationService;
@@ -31,22 +30,12 @@ public class ReservationEventsConsumer {
 			String eventType = debeziumEventParser.getEventType(message);
 
 			switch (EventType.from(eventType)) {
-				case RESERVATION_CONFIRM_REQUESTED -> {
-					EventEnvelope<PaymentEvent.PaymentCompletedEvent> envelope =
-						debeziumEventParser.parse(message, PaymentEvent.PaymentCompletedEvent.class);
-					reservationService.confirmReservation(envelope.payload());
-				}
 				case RESERVATION_CONFIRMED -> {
 					EventEnvelope<ReservationEvent.ReservationConfirmedEvent> envelope =
 						debeziumEventParser.parse(message, ReservationEvent.ReservationConfirmedEvent.class);
 					ReservationEvent.ReservationConfirmedEvent event = envelope.payload();
 					reservationHoldService.removeHold(event.accommodationId(), event.checkInDate(), event.checkOutDate());
 					log.info("[KAFKA] 예약 확정 완료. Redis 홀드 제거. Accommodation ID={}", event.accommodationId());
-				}
-				case RESERVATION_EXPIRE_REQUESTED -> {
-					EventEnvelope<PaymentEvent.PaymentFailedEvent> envelope =
-						debeziumEventParser.parse(message, PaymentEvent.PaymentFailedEvent.class);
-					reservationService.expireReservation(envelope.payload());
 				}
 				case RESERVATION_EXPIRED -> {
 					EventEnvelope<ReservationEvent.ReservationExpiredEvent> envelope =
