+++ src/main/java/kr/kro/airbob/outbox/EventType.java
@@ -10,9 +10,7 @@ import lombok.RequiredArgsConstructor;
 public enum EventType {
 
 	// 예약 이벤트
-	RESERVATION_CONFIRM_REQUESTED("RESERVATION", "reservation-events"), // 예약 확정 요청
 	RESERVATION_CONFIRMED("RESERVATION", "reservation-events"), // 예약 확정 완료
-	RESERVATION_EXPIRE_REQUESTED("RESERVATION", "reservation-events"), // 예약 만료 요청
 	RESERVATION_EXPIRED("RESERVATION", "reservation-events"),   // 예약 만료 완료
 	RESERVATION_PENDING("RESERVATION", "reservation-events"), // 예약 보류
 	@Deprecated
@@ -23,16 +21,9 @@ public enum EventType {
 
 	// 결제 이벤트
 	PAYMENT_EXECUTION_REQUESTED_V1("PAYMENT_OPERATION", "PAYMENT_OPERATION.events"),
-	PAYMENT_CONFIRM_REQUESTED("PAYMENT", "payment-events"), // 결제 승인 요청
-	PAYMENT_COMPLETED("PAYMENT", "payment-events"), // 결제 완료
 	PAYMENT_CANCELLATION_REQUESTED("PAYMENT", "payment-events"), // 결제 취소 요청
 	PAYMENT_CANCELLATION_COMPLETED("PAYMENT", "payment-events"), // 결제 취소 완료
-	PAYMENT_FAILED("PAYMENT", "payment-events"),
 	PAYMENT_CANCELLATION_FAILED("PAYMENT", "payment-events"),
-
-	PG_CALL_REQUESTED("PAYMENT", "payment-events"),       // PG API 호출 요청
-	PG_CALL_SUCCEEDED("PAYMENT", "payment-events"),       // PG API 호출 성공
-	PG_CALL_FAILED("PAYMENT", "payment-events"), // PG API 호출 실패
 
 	PG_CANCEL_CALL_REQUESTED("PAYMENT", "payment-events"),   // PG 취소 API 호출 요청
 	PG_CANCEL_CALL_SUCCEEDED("PAYMENT", "payment-events"),   // PG 취소 API 호출 성공
