+++ src/test/java/kr/kro/airbob/domain/reservation/service/ReservationServiceTest.java
@@ -26,7 +26,6 @@ import kr.kro.airbob.domain.accommodation.repository.AccommodationRepository;
 import kr.kro.airbob.domain.accommodation.repository.projection.AccommodationBookingProjection;
 import kr.kro.airbob.domain.member.entity.Member;
 import kr.kro.airbob.domain.payment.dto.PaymentRequest;
-import kr.kro.airbob.domain.payment.event.PaymentEvent;
 import kr.kro.airbob.domain.reservation.dto.ReservationRequest;
 import kr.kro.airbob.domain.reservation.dto.ReservationResponse;
 import kr.kro.airbob.domain.reservation.entity.Reservation;
@@ -360,47 +359,6 @@ class ReservationServiceTest {
 
 			// then
 			then(transactionService).should().cancelReservationInTx(reservationUid, cancelRequest, memberId);
-		}
-	}
-
-	@Nested
-	@DisplayName("예약 확정 테스트")
-	class ConfirmReservationTest {
-
-		@Test
-		@DisplayName("결제 완료 이벤트 수신 시 transactionService에 위임된다")
-		void 정상_확정_위임() {
-			// given
-			String reservationUid = UUID.randomUUID().toString();
-			PaymentEvent.PaymentCompletedEvent event = new PaymentEvent.PaymentCompletedEvent(reservationUid);
-
-			// when
-			reservationService.confirmReservation(event);
-
-			// then
-			then(transactionService).should().confirmReservationInTx(reservationUid);
-		}
-	}
-
-	@Nested
-	@DisplayName("예약 만료 테스트")
-	class ExpireReservationTest {
-
-		@Test
-		@DisplayName("결제 실패 이벤트 수신 시 transactionService에 위임된다")
-		void 정상_만료_위임() {
-			// given
-			String reservationUid = UUID.randomUUID().toString();
-			PaymentEvent.PaymentFailedEvent event = new PaymentEvent.PaymentFailedEvent(
-				reservationUid,
-				"결제 시간 초과"
-			);
-
-			// when
-			reservationService.expireReservation(event);
-
-			// then
-			then(transactionService).should().expireReservationInTx(reservationUid, "결제 시간 초과");
 		}
 	}
