+++ src/test/java/kr/kro/airbob/domain/reservation/service/ReservationTransactionServiceTest.java
@@ -46,7 +46,6 @@ import kr.kro.airbob.domain.reservation.entity.ReservationHistory;
 import kr.kro.airbob.domain.reservation.event.ReservationEvent;
 import kr.kro.airbob.domain.reservation.exception.ReservationAccessDeniedException;
 import kr.kro.airbob.domain.reservation.exception.ReservationConflictException;
-import kr.kro.airbob.domain.reservation.exception.ExpiredReservationConfirmationException;
 import kr.kro.airbob.domain.reservation.exception.InvalidReservationDateException;
 import kr.kro.airbob.domain.reservation.exception.ReservationNotFoundException;
 import kr.kro.airbob.domain.reservation.exception.ReservationOutsideBookingWindowException;
@@ -468,317 +467,6 @@ class ReservationTransactionServiceTest {
 		}
 	}
 
-	@Nested
-	@DisplayName("예약 확정 트랜잭션 테스트")
-	class ConfirmReservationInTxTest {
-
-		@Test
-		@DisplayName("PAYMENT_PROCESSING 상태에서 확정 시 CONFIRMED로 변경된다")
-		void 정상_확정() {
-			// given
-			UUID reservationUid = UUID.randomUUID();
-			Reservation reservation = createReservationWithStatus(
-				reservationUid, ReservationStatus.PAYMENT_PROCESSING);
-			Payment payment = mock(Payment.class);
-			given(payment.getApprovedAt()).willReturn(NOW);
-
-			given(reservationRepository.findByReservationUidWithLock(reservationUid))
-				.willReturn(Optional.of(reservation));
-			given(paymentRepository.findByReservationReservationUidWithLock(reservationUid))
-				.willReturn(Optional.of(payment));
-
-			// when
-			transactionService.confirmReservationInTx(reservationUid.toString());
-
-			// then
-			assertThat(reservation.getStatus()).isEqualTo(ReservationStatus.CONFIRMED);
-
-			then(historyRepository).should().save(historyCaptor.capture());
-			ReservationHistory savedHistory = historyCaptor.getValue();
-			assertThat(savedHistory.getStatus()).isEqualTo(ReservationStatus.CONFIRMED);
-			assertThat(savedHistory.getChangeType()).isEqualTo(ChangeType.STATUS_CHANGE);
-			assertThat(savedHistory.getSourceSystem()).isEqualTo("KAFKA");
-
-			then(outboxEventPublisher).should().save(eq(EventType.RESERVATION_CONFIRMED), any(ReservationEvent.ReservationConfirmedEvent.class));
-		}
-
-		@Test
-		@DisplayName("이벤트 처리가 늦어도 PG 승인 시각이 만료 전이면 예약을 확정한다")
-		void confirmsUsingPgApprovalTimeInsteadOfEventProcessingTime() {
-			UUID reservationUid = UUID.randomUUID();
-			Instant expiresAt = NOW.minusSeconds(5);
-			Reservation reservation = Reservation.builder()
-				.id(1L)
-				.reservationUid(reservationUid)
-				.reservationCode("ABC123")
-				.accommodation(accommodation)
-				.guest(guest)
-				.checkInDate(LocalDate.of(2026, 8, 15))
-				.checkOutDate(LocalDate.of(2026, 8, 16))
-				.checkInAt(Instant.parse("2026-08-15T19:00:00Z"))
-				.checkOutAt(Instant.parse("2026-08-16T15:00:00Z"))
-				.timeZoneId(TIME_ZONE_ID)
-				.guestCount(2)
-				.totalPrice(100_000L)
-				.currency("KRW")
-				.status(ReservationStatus.PAYMENT_PROCESSING)
-				.expiresAt(expiresAt)
-				.build();
-			Payment payment = mock(Payment.class);
-			given(payment.getApprovedAt()).willReturn(expiresAt.minusNanos(1));
-			given(reservationRepository.findByReservationUidWithLock(reservationUid))
-				.willReturn(Optional.of(reservation));
-			given(paymentRepository.findByReservationReservationUidWithLock(reservationUid))
-				.willReturn(Optional.of(payment));
-
-			transactionService.confirmReservationInTx(reservationUid.toString());
-
-			assertThat(reservation.getStatus()).isEqualTo(ReservationStatus.CONFIRMED);
-		}
-
-		@Test
-		@DisplayName("이미 CONFIRMED 상태면 조기 반환한다 (멱등성)")
-		void 멱등성_이미_확정() {
-			// given
-			UUID reservationUid = UUID.randomUUID();
-			Reservation reservation = createReservationWithStatus(reservationUid, ReservationStatus.CONFIRMED);
-
-			given(reservationRepository.findByReservationUidWithLock(reservationUid))
-				.willReturn(Optional.of(reservation));
-
-			// when
-			transactionService.confirmReservationInTx(reservationUid.toString());
-
-			// then
-			assertThat(reservation.getStatus()).isEqualTo(ReservationStatus.CONFIRMED);
-			then(historyRepository).should(never()).save(any());
-			then(outboxEventPublisher).should(never()).save(any(), any());
-		}
-
-		@Test
-		@DisplayName("확정 후 취소 처리 중인 예약에 중복 결제 완료 이벤트가 와도 보상 대상으로 만들지 않는다")
-		void skipsDuplicateConfirmationAfterCancellationStarted() {
-			UUID reservationUid = UUID.randomUUID();
-			Reservation reservation = createReservationWithStatus(
-				reservationUid, ReservationStatus.CANCELLATION_FAILED);
-			given(reservationRepository.findByReservationUidWithLock(reservationUid))
-				.willReturn(Optional.of(reservation));
-
-			transactionService.confirmReservationInTx(reservationUid.toString());
-
-			assertThat(reservation.getStatus()).isEqualTo(ReservationStatus.CANCELLATION_FAILED);
-			then(historyRepository).shouldHaveNoInteractions();
-			then(outboxEventPublisher).shouldHaveNoInteractions();
-		}
-
-		@Test
-		@DisplayName("결제 만료 시각과 정확히 같으면 예약을 확정하지 않는다")
-		void rejectsConfirmationAtExactExpiryInstant() {
-			UUID reservationUid = UUID.randomUUID();
-			Reservation reservation = Reservation.builder()
-				.id(1L)
-				.reservationUid(reservationUid)
-				.reservationCode("ABC123")
-				.accommodation(accommodation)
-				.guest(guest)
-				.checkInDate(LocalDate.of(2026, 8, 15))
-				.checkOutDate(LocalDate.of(2026, 8, 16))
-				.checkInAt(Instant.parse("2026-08-15T19:00:00Z"))
-				.checkOutAt(Instant.parse("2026-08-16T15:00:00Z"))
-				.timeZoneId(TIME_ZONE_ID)
-				.guestCount(2)
-				.totalPrice(100_000L)
-				.currency("KRW")
-				.status(ReservationStatus.PAYMENT_PROCESSING)
-				.expiresAt(NOW)
-				.build();
-			Payment payment = mock(Payment.class);
-			given(payment.getApprovedAt()).willReturn(NOW);
-			given(reservationRepository.findByReservationUidWithLock(reservationUid))
-				.willReturn(Optional.of(reservation));
-			given(paymentRepository.findByReservationReservationUidWithLock(reservationUid))
-				.willReturn(Optional.of(payment));
-
-			assertThatThrownBy(() -> transactionService.confirmReservationInTx(reservationUid.toString()))
-				.isInstanceOf(ExpiredReservationConfirmationException.class);
-
-			assertThat(reservation.getStatus()).isEqualTo(ReservationStatus.PAYMENT_PROCESSING);
-			then(historyRepository).shouldHaveNoInteractions();
-			then(outboxEventPublisher).shouldHaveNoInteractions();
-		}
-
-		@Test
-		@DisplayName("결제 선점 없이 PAYMENT_PENDING인 예약은 확정하지 않는다")
-		void rejectsConfirmationWithoutPaymentClaim() {
-			UUID reservationUid = UUID.randomUUID();
-			Reservation reservation = createReservationWithStatus(
-				reservationUid, ReservationStatus.PAYMENT_PENDING);
-			given(reservationRepository.findByReservationUidWithLock(reservationUid))
-				.willReturn(Optional.of(reservation));
-
-			assertThatThrownBy(() -> transactionService.confirmReservationInTx(reservationUid.toString()))
-				.isInstanceOf(ExpiredReservationConfirmationException.class);
-
-			assertThat(reservation.getStatus()).isEqualTo(ReservationStatus.PAYMENT_PENDING);
-			then(paymentRepository).shouldHaveNoInteractions();
-			then(historyRepository).shouldHaveNoInteractions();
-		}
-
-		@Test
-		@DisplayName("예약이 존재하지 않으면 ReservationNotFoundException이 발생한다")
-		void 예외_예약_미존재() {
-			// given
-			UUID reservationUid = UUID.randomUUID();
-			given(reservationRepository.findByReservationUidWithLock(reservationUid))
-				.willReturn(Optional.empty());
-
-			// when & then
-			assertThatThrownBy(() -> transactionService.confirmReservationInTx(reservationUid.toString()))
-				.isInstanceOf(ReservationNotFoundException.class);
-		}
-	}
-
-	@Nested
-	@DisplayName("예약 만료 트랜잭션 테스트")
-	class ExpireReservationInTxTest {
-
-		@Test
-		@DisplayName("PAYMENT_PENDING 상태에서 만료 시 EXPIRED로 변경된다")
-		void 정상_만료() {
-			// given
-			UUID reservationUid = UUID.randomUUID();
-			Reservation reservation = createReservationWithStatus(reservationUid, ReservationStatus.PAYMENT_PENDING);
-
-			given(reservationRepository.findByReservationUidWithLock(reservationUid))
-				.willReturn(Optional.of(reservation));
-
-			// when
-			transactionService.expireReservationInTx(reservationUid.toString(), "결제 시간 초과");
-
-			// then
-			assertThat(reservation.getStatus()).isEqualTo(ReservationStatus.EXPIRED);
-
-			then(historyRepository).should().save(historyCaptor.capture());
-			ReservationHistory savedHistory = historyCaptor.getValue();
-			assertThat(savedHistory.getStatus()).isEqualTo(ReservationStatus.EXPIRED);
-			assertThat(savedHistory.getChangeType()).isEqualTo(ChangeType.STATUS_CHANGE);
-			assertThat(savedHistory.getSourceSystem()).isEqualTo("KAFKA");
-
-			then(outboxEventPublisher).should().save(eq(EventType.RESERVATION_EXPIRED), any(ReservationEvent.ReservationExpiredEvent.class));
-		}
-
-		@Test
-		@DisplayName("이미 EXPIRED 상태면 조기 반환한다 (멱등성)")
-		void 멱등성_이미_만료() {
-			// given
-			UUID reservationUid = UUID.randomUUID();
-			Reservation reservation = createReservationWithStatus(reservationUid, ReservationStatus.EXPIRED);
-
-			given(reservationRepository.findByReservationUidWithLock(reservationUid))
-				.willReturn(Optional.of(reservation));
-
-			// when
-			transactionService.expireReservationInTx(reservationUid.toString(), "결제 시간 초과");
-
-			// then
-			assertThat(reservation.getStatus()).isEqualTo(ReservationStatus.EXPIRED);
-			then(historyRepository).should(never()).save(any());
-			then(outboxEventPublisher).should(never()).save(any(), any());
-		}
-
-		@Test
-		@DisplayName("예약이 존재하지 않으면 ReservationNotFoundException이 발생한다")
-		void 예외_예약_미존재() {
-			// given
-			UUID reservationUid = UUID.randomUUID();
-			given(reservationRepository.findByReservationUidWithLock(reservationUid))
-				.willReturn(Optional.empty());
-
-			// when & then
-			assertThatThrownBy(() -> transactionService.expireReservationInTx(reservationUid.toString(), "결제 시간 초과"))
-				.isInstanceOf(ReservationNotFoundException.class);
-		}
-	}
-
-	@Nested
-	@DisplayName("결제 보상 준비 트랜잭션 테스트")
-	class PreparePaymentCompensationInTxTest {
-
-		@Test
-		@DisplayName("PAYMENT_PENDING 예약은 잠근 뒤 EXPIRED로 전환해 결제 보상을 허용한다")
-		void pendingReservationIsFencedBeforeCompensation() {
-			UUID reservationUid = UUID.randomUUID();
-			Reservation reservation = createReservationWithStatus(
-				reservationUid, ReservationStatus.PAYMENT_PENDING);
-			given(reservationRepository.findByReservationUidWithLock(reservationUid))
-				.willReturn(Optional.of(reservation));
-
-			boolean shouldCompensate = transactionService.preparePaymentCompensationInTx(
-				reservationUid.toString(), "예약 확정 최종 실패");
-
-			assertThat(shouldCompensate).isTrue();
-			assertThat(reservation.getStatus()).isEqualTo(ReservationStatus.EXPIRED);
-			then(historyRepository).should().save(any(ReservationHistory.class));
-			then(outboxEventPublisher).should().save(
-				eq(EventType.RESERVATION_EXPIRED),
-					any(ReservationEvent.ReservationExpiredEvent.class));
-		}
-
-		@Test
-		@DisplayName("PAYMENT_PROCESSING 예약도 잠근 뒤 EXPIRED로 전환해 결제 보상을 허용한다")
-		void processingReservationIsFencedBeforeCompensation() {
-			UUID reservationUid = UUID.randomUUID();
-			Reservation reservation = createReservationWithStatus(
-				reservationUid, ReservationStatus.PAYMENT_PROCESSING);
-			given(reservationRepository.findByReservationUidWithLock(reservationUid))
-				.willReturn(Optional.of(reservation));
-
-			boolean shouldCompensate = transactionService.preparePaymentCompensationInTx(
-				reservationUid.toString(), "승인 시각 만료");
-
-			assertThat(shouldCompensate).isTrue();
-			assertThat(reservation.getStatus()).isEqualTo(ReservationStatus.EXPIRED);
-			then(historyRepository).should().save(any(ReservationHistory.class));
-			then(outboxEventPublisher).should().save(
-				eq(EventType.RESERVATION_EXPIRED),
-				any(ReservationEvent.ReservationExpiredEvent.class));
-		}
-
-		@Test
-		@DisplayName("이미 CONFIRMED인 예약은 결제 보상을 허용하지 않는다")
-		void confirmedReservationCannotBeCompensated() {
-			UUID reservationUid = UUID.randomUUID();
-			Reservation reservation = createReservationWithStatus(
-				reservationUid, ReservationStatus.CONFIRMED);
-			given(reservationRepository.findByReservationUidWithLock(reservationUid))
-				.willReturn(Optional.of(reservation));
-
-			boolean shouldCompensate = transactionService.preparePaymentCompensationInTx(
-				reservationUid.toString(), "예약 확정 최종 실패");
-
-			assertThat(shouldCompensate).isFalse();
-			assertThat(reservation.getStatus()).isEqualTo(ReservationStatus.CONFIRMED);
-			then(historyRepository).shouldHaveNoInteractions();
-			then(outboxEventPublisher).shouldHaveNoInteractions();
-		}
-
-		@Test
-		@DisplayName("이미 EXPIRED인 예약은 결제 보상을 계속 허용한다")
-		void expiredReservationCanBeCompensatedIdempotently() {
-			UUID reservationUid = UUID.randomUUID();
-			Reservation reservation = createReservationWithStatus(
-				reservationUid, ReservationStatus.EXPIRED);
-			given(reservationRepository.findByReservationUidWithLock(reservationUid))
-				.willReturn(Optional.of(reservation));
-
-			boolean shouldCompensate = transactionService.preparePaymentCompensationInTx(
-				reservationUid.toString(), "예약 확정 최종 실패");
-
-			assertThat(shouldCompensate).isTrue();
-			then(historyRepository).shouldHaveNoInteractions();
-			then(outboxEventPublisher).shouldHaveNoInteractions();
-		}
-	}
 
 	@Nested
 	@DisplayName("예약 취소 트랜잭션 테스트")
